DB기능 수정 및 내용 추가

+++ RootTools/Database/Table/DatabaseManager.cs
@@ -224,8 +224,11 @@ namespace RootTools.Database
 
 		public void ClearTableData(string _sTable)
 		{
-			string sTableClearQuery = "truncate " + _sTable;
-			DatabaseManager.Instance.SendQuery(sTableClearQuery);
+			if(CheckExistTable(_sTable))
+			{
+				string sTableClearQuery = "truncate " + _sTable;
+				DatabaseManager.Instance.SendQuery(sTableClearQuery);
+			}
 		}
 
 
@@ -353,14 +356,24 @@ namespace RootTools.Database
 		public bool CheckExistTable(string tableName)
 		{
 			string query = string.Format("SELECT * FROM {0};", tableName);
-			var code = SendQuery(query);
-			if(code == (int)DB_ERROR.TABLE_IS_MISSING)
+			try
 			{
-				return false;
+				var code = SendQuery(query);
+				if (code == 0)
+					return true;//다른 에러일 수도 있으니 예외처리등이 필요? ㅁ?ㄹ
+				else
+					return false;//추가 예외처리 필요
 			}
-			else
+			catch (MySqlException ex)
 			{
-				return true;//다른 에러일 수도 있으니 예외처리등이 필요? ㅁ?ㄹ
+				if (ex.Code == (int)DB_ERROR.TABLE_IS_MISSING)
+				{
+					return false;
+				}
+				else
+				{
+					return false;//추가 예외처리 필요
+				}
 			}
 		}
 		public void CreateTable(string tableName, Type type, string keyName)
@@ -374,25 +387,21 @@ namespace RootTools.Database
 				string nullAble = " NULL";
 				if(keyName == fld[i].Name)
 				{
-					nullAble = " NOT NULL";
-					if(fld[i].FieldType == typeof(int))
-					{
-						nullAble += " AUTO_INCREMENT";
-					}
+					nullAble = " NOT NULL AUTO_INCREMENT";
 				}
 				nullAble += ",";
 
 				if (fld[i].FieldType == typeof(int))
 				{
-					stbr.Append(" int(11) DEFAULT"+ nullAble);
+					stbr.Append(" int(11) " + nullAble);
 				}
 				else if (fld[i].FieldType == typeof(float))
 				{
-					stbr.Append(" double DEFAULT"+ nullAble);
+					stbr.Append(" double "+ nullAble);
 				}
 				else if (fld[i].FieldType == typeof(string) || fld[i].FieldType == typeof(String))
 				{
-					stbr.Append(" varchar(45) DEFAULT"+ nullAble);
+					stbr.Append(" varchar(45) "+ nullAble);
 				}
 			}
 			stbr.Append(string.Format("PRIMARY KEY ({0})) ENGINE=InnoDB DEFAULT CHARSET=utf8;", keyName));
@@ -410,7 +419,7 @@ namespace RootTools.Database
 					//var tableQuery = string.Format("CREATE TABLE '{0}' ('m_nDefectIndex' int(11) NOT NULL AUTO_INCREMENT,'m_strInspectionID' varchar(45) DEFAULT NULL,'m_nDefectCode' int(11) DEFAULT NULL,'m_fSize' double DEFAULT NULL,'m_fWidth' double DEFAULT NULL,'m_fHeight' double DEFAULT NULL,'m_fRelX' double DEFAULT NULL,'m_fRelY' double DEFAULT NULL,'m_fAbsX' double DEFAULT NULL,'m_fAbsY' double DEFAULT NULL,'m_fGV' double DEFAULT NULL,'m_nChipIndexX' int(11) DEFAULT NULL,'m_nCHipIndexY' int(11) DEFAULT NULL,PRIMARY KEY ('m_nDefectIndex')) ENGINE=InnoDB AUTO_INCREMENT=7 DEFAULT CHARSET=utf8;", tableName);
 					CreateTable(tableName, typeof(Defect),nameof(Defect.m_nDefectIndex));
 				}
-				SendQuery(string.Format("TRUNCATE {0};", tableName));
+				//SendQuery(string.Format("TRUNCATE {0};", tableName));
 				StringBuilder temp = new StringBuilder();
 				StringBuilder sbQuery = new StringBuilder();
 				StringBuilder sbColumList = new StringBuilder();
